KA-161 Refactor initial state
